+++ src/main/java/znu/visum/components/movies/usecases/removefavorite/domain/RemoveFavoriteMovieService.java
@@ -1,0 +1,5 @@
+package znu.visum.components.movies.usecases.removefavorite.domain;
+
+public interface RemoveFavoriteMovieService {
+  boolean removeFavorite(long movieId);
+}